Split off rational_basic in 2 separate tests.
