+++ test/CMakeLists.txt
@@ -126,10 +126,8 @@ ADD_MPPP_TESTCASE(rational_arith)
 ADD_MPPP_TESTCASE(rational_arith_ops_01)
 ADD_MPPP_TESTCASE(rational_arith_ops_02)
 ADD_MPPP_TESTCASE(rational_arith_ops_03)
-if(NOT MINGW)
-    # NOTE: same as above.
-    ADD_MPPP_TESTCASE(rational_basic)
-endif()
+ADD_MPPP_TESTCASE(rational_basic_01)
+ADD_MPPP_TESTCASE(rational_basic_02)
 ADD_MPPP_TESTCASE(rational_binomial)
 ADD_MPPP_TESTCASE(rational_hash)
 ADD_MPPP_TESTCASE(rational_inv)
